Updated bootstrap to clone or pull vundle
Vundle was unable to manage itself using the previous version of this
script. I'm guessing that it's because the .git directory in the
vundle dir wasn't being copied across, which meant that vundle wasn't
seeing itself as a valid plugin.

Deleted vundle plugin from repo, as it is being installed by the
bootstrap script.