Wrap autocmd blocks with if and augroup
Old versions of vim don't have autocmd, and these calls would cause
lots of shouting.

Wrapping in groups means that we are able to source vimrc without
adding new autocmds. This leads to slow downs and potential weirdness.
This was probably where the issue I was having with saving vim (which
triggers a sourcing of the vimrc) to be really slow and give me black
screens for a LOOOONG time.