Switched from Vundle to vim-plug
https://github.com/junegunn/vim-plug

Primarily for its post-update hooks, meaning that it's no longer
necessrary to compile modules in bootstrap, which simplifies the
bootstrap file a lot.

Also good for on-demand loading, because I don't want to have NERDTree
clogging up memory everytime I load vim, it should only be loaded when I
actually use it (which rarely).

Oh, and parallel installationa and update.